+++ tests/testthat/test03-HLAAllele.R
@@ -1,56 +1,80 @@
 context("HLAAlleles")
 
-doc <- read_hla_xml(remote = FALSE)
-test_that("read_hla_xml returns correctly", {
+doc <- if (remote_tests) {
+  read_hla_xml(db_path = tmp, remote = FALSE)
+} else if (local_repo) {
+  read_hla_xml(db_path = getOption("hlatools.local_repos"), remote = FALSE)
+}
+
+test_that("read_hla_xml() works", {
   expect_is(doc, "xml_document")
 })
 
 dpa <- parse_hla_alleles(doc, "HLA-DPA1")
-test_that("parse_hla_alleles returns correctly", {
+
+test_that("HLAAllele global API", {
   expect_is(dpa, "HLAAllele")
+  ##
+  expect_error(db_version(dpa))
+  expect_error(hlatools_version(dpa))
   expect_equal(locusname(dpa), "HLA-DPA1")
+  ##
+  ## Features
+  expect_is(features(dpa), "CompressedHLARangesList")
+  expect_is(features(dpa)[[1]], "HLARanges")
+  expect_equal(length(features(dpa)[[1]]), 9)
+  ## Sequences
+  expect_is(sequences(dpa), "DNAStringSet")
+  expect_equal(length(sequences(dpa)), 64)
+  expect_equal(names(sequences(dpa))[[1]], "HLA-DPA1*01:03:01:01")
+  expect_equal(width(sequences(dpa))[[1]], 9775)
+  ## Metadata
+  expect_is(elementMetadata(dpa), "DataFrame")
+  expect_equal(length(elementMetadata(dpa)), 14)
 })
 
-test_that("HLAAllele accessor methods work correctly", {
+test_that("HLAAllele element-wise API", {
   ## Subsetting
-  x <- dpa["01:03:01:01"]
+  x <- dpa["HLA-DPA1*01:03:01:01"]
   expect_equal(length(x), 1)
+  ##
+  expect_equal(allele_id(x), "HLA00499")
+  ##
   expect_equal(allele_name(x), "HLA-DPA1*01:03:01:01")
   expect_equal(names(x), "HLA-DPA1*01:03:01:01")
-  expect_equal(allele_id(x), "HLA00499")
+  ##
   expect_equal(g_group(x), "DPA1*01:03:01G")
+  ##
   expect_equal(p_group(x), "DPA1*01:03P")
+  ##
   expect_equal(cwd_status(x), "Common")
+  ##
   expect_equal(ethnicity(x), "Caucasoid:Oriental")
+  ##
   expect_equal(sample_name(x), "BOLETH:JM15:LB:LG2:PRIESS:QBL:T5-1:TUBO:WJR076")
+  ##
   expect_true(is_complete(x))
   expect_false(is_lsl(x))
+  ##
   expect_is(as(x, "data.table"), "tbl_dt")
-  ## Partial subsettig
-  y <- dpa["01:03"]
-  expect_equal(length(y), 15)
-  ## Combine
-  expect_equal(length(c(x, y)), 16)
-
 })
 
-test_that("HLAAllele features, sequences, and metadata", {
-  x <- dpa["01:03:01:01"]
-  ## Features
-  f <- features(x)
-  expect_is(f, "CompressedHLARangesList")
-  expect_is(f[[1]], "HLARanges")
-  expect_equal(length(f[[1]]), 9)
-  ## Sequences
-  s <- sequences(x)
-  expect_is(s, "DNAStringSet")
-  expect_equal(length(s), 1)
-  expect_equal(names(s), "HLA-DPA1*01:03:01:01")
-  expect_equal(width(s), 9775)
-  ## Metadata
-  m <- elementMetadata(x)
-  expect_is(m, "DataFrame")
-  expect_equal(length(m), 14)
+test_that("HLAAllele subsetting", {
+  ## Subsetting by allele name
+  expect_equal(NROW(dpa["HLA-DPA1*01:03:01:01"]), 1)
+  expect_equal(NROW(dpa["DPA1*01:03:01:01"]), 1)
+  expect_equal(NROW(dpa["01:03:01:01"]), 1)
+  ## Partial subsetting by allele name
+  expect_equal(NROW(dpa["01:03:01"]), 11)
+  expect_equal(NROW(dpa["DPA1*01:03"]), 15)
+  expect_equal(NROW(dpa["HLA-DPA1*01"]), 27)
+  ## Subsetting by number
+  expect_equal(NROW(dpa[1]), 1)
+  expect_equal(NROW(dpa[1:4]), 4)
+  ## Combine
+  x <- dpa[1:2]
+  y <- dpa[3:4]
+  expect_equal(NROW(c(x, y)), 4)
 })
 
 test_that("HLAAllele exon, intron, and utr", {
@@ -71,5 +95,3 @@ test_that("HLAAllele exon, intron, and utr", {
   expect_equal(Biostrings::toString(Biostrings::subseq(u, 1, 10)), "AATACTGTAA")
   expect_error(utr(x, 3))
 })
-
-